[KAI HTTP Client Method]: Create & Test "booking_with_arv_info"

+++ kaiwsdkv2/response_types_test.go
@@ -538,7 +538,7 @@ func TestBookingWithArvInfoRSOK(t *testing.T) {
 	arriveTime := "0538"
 	trainNo := "44"
 
-	var adultCount float64 = 4
+	var adultCount int = 4
 	paxName0 := "ARGO PARAHYANGAN"
 	wagonCode := "EKS"
